feat(highlights): add ArticleComponentHighlights and move out code from SavedItemViewModel
